WIP: Split up app_config.dart
- Create file_service.dart to handle loading a JSON file.
- Create app_site_config.dart to hold the AppSiteConfig value object.
