+++ lib/values/app_site_config.dart
@@ -1,0 +1,11 @@
+class AppSiteConfig {
+  AppSiteConfig({this.hosts});
+
+  static AppSiteConfig fromJson(Map<String, dynamic> json) {
+    return AppSiteConfig(
+      hosts: json != null ? json['hosts']?.cast<String>() : null,
+    );
+  }
+
+  final List<String> hosts;
+}
